Collection.until() method
Added Collection.until() method that works similar to limit() but rather
than specifying a number, you specify a filter function that returns
true when it's time to stop iteration.

Also, added the reverse() method to Table.

Also, removed the previous autocomplete fix because it destroyed
autocomplete rather than repairing it.