Dexie.Syncable version 0.1 (not tested yet)
The Dexie.Syncable addon enables two-way syncronization with a remote
server. The addon is dependant on Dexie.Observable and a custom
implementation of the ISyncProtocol API defined in
Dexie.Syncable.SyncProtocolAPI.js.

There are examples in
sampes/remote-sync/websocket/WebSocketSyncProtocol.js and
WebSocketSyncServer.js. Currently these examples are not yet tested. The
aim is to create a unit test using these samples and verifying the
entire sync framework.