Differentiating syncedRevision from baseRevision
After a review of the sync API I found out that there is a difference
between baseRevision and syncedRevision.

BaseRevision is what the given changes are based on while syncedRevision
is the revision that we request the server to give changes from.

This commit updates the API spec, the sample code and the framework
(Dexie.Syncable.js) to differentiate these.