Bug #11: anyOf() fails when comparing integers
Was using Array.sort() which actually sorts integers by converting them
to strings first. Changed to use sort() by comparing less than / greater
than (which works as expected for strings and Numbers). Could be calling
indexedDB.cmp() (which is probably designed for this purpose), but the
disadvantage with that one is that is throws whenever reaching a null
value. Need to have a sorter that doesnt throw and it I'm not fond of
doing try/catch in in the sorter due to performance reasons.