Unit test triggered a Firefox stalling bug
In one unit test, a transaction was opened against two instances of the
users table.  Firefox got into an eternal stall state when having called
IDBDatabase.transaction() with the same store name occurring twice in
the storeNames argument.

Also fixed in Dexie.js that it will throw an error if caller specifies
the same table twice.

Another bugfix: Promise implementation could throw in case it was
resolved at first and then rejected. Behavior was still as expected
earlier but with an unnescessary exception being thrown.