Bugfix: non-inbound keys didnt work
All samples and unit tests so far in Dexie has been done using inbound
primary key ('++id' for example). When testing to omit the name of the
primary key (as I have documented in wiki) I detected bugs in the code
related to this. This checkin should solve those bugs.