Database Monitoring capability added
Possible to monitor databases on current origin when not knowing
anything about what databases that exists or their schemas.

Features that enables this:
1. Static method: Dexie.getDatabaseNames() - retrieve a list of
databases at current host.
2. Possible to call db.open() when not having specified any versions or
schemas. Database will then open any existing database and if no
database exists it will fail.
3. New property db.verno tells current version of the database.

Together this makes it possible to create a DB monitor tool that would
look up any databases on the current origin and display the schemas for
each of them as well as display or modify their contents.

(It also enables non-Dexie IDBs to be opened by Dexie without having to
translate the existing schema into a dexie format)

Other changes in this commit:
* Removed FireFox workaround since the situation is very rare and the
bug has been resolved in Firefox, so it wont be a problem in future.