Debugging improvement: reject Error keep orig
When onerror was fired, we used to derive the original Error object to a
simple [toString: function (){...}] object that would add some info to
the orignal object. What I found was that this hides the actual Error
object in the debugger so that you have to call toString() in the
debugger  watchers in order to see what went wrong. Now I changed this
so that the same Error instance is used but its message is extended with
the additional info.