Dexie.Syncable.js: Days of testing and bugfixing
Tested Dexie.Syncable by creating a sync provider towards the sync
server of my current project. The goal is to migrate my app from using
localStorage to be using indexedDB. I will replace the current db
framework built on localStorage and ajax syncing to be based on Dexie
and Dexie.Syncable.

When committing this, the framework is working as expected with my
application. The application runs in two different windows so one of the
windows becomes the master node that is responsible of doing the sync
calls to the server. There were errors in forms of dead-locks due to the
lack of db.vip() in some places.

Still, one of the unit tests for Dexie.Syncable fails. Need to check
what is wrong and do another commit soon.