db.vip() prohibit deadlock for on('ready') subscribers
A new method db.vip() is needed when subscribing to db.on('ready') and
returning a Promise; Returning a Promise from on db.on('ready') makes
the database stay blocked until the returned Promise is resolved. If the
subscriber then want to use database during this time, it's calls will
also be blocked, resulting in a deadlock. However, thanks to the magic
Promise.PSD.letThrough boolean, this deadlock never happens if the
subscriber is consequently using Dexie Promises to fullfill any async
job, but sometimes that is not possible. For example, the each() method
does not execute in the context of a Promise. Nor does any external
event or ajax call. In those cases, the problem can be worked around by
surrounding any database request with the db.vip() method.