Fine tuning the sample
The sample now dumps the database schemas in form of javascript code;
the code that you would do to define the databases with Dexie. If you
run the sample on Opera or Chrome on an indexedDB database not created
by Dexie, you will see how you should define the database with Dexie if
you plan to migrate to Dexie.