Dexie.Syncable: Corrected merge algorithm
When merging two update-changes, we didnt handle the case when the
previous change contained a parent key to the next change's keyPath.
What would happen before is that the merged change would contain both
the parent keyPath and the sub keyPath. When this change would have been
applied, it would still be OK if we were certain that the parent path
would apply first and sub-path second, but that's not how it is
implemented.

To canonicalize the change, I added a check for parent path and if it
exists, let the new change apply to the contents of the prev change
instead.