New transact() method merged with transaction()
The existing transaction() method now works as the new transact()
method. This will make migrating to the new way of handling transactions
easier.

A difference with db.transaction() is now that instead of throwing when
user tries to access db.[table] directly in the transaction scope, we
now allow it. In fact I plan to change the documentation to promote this
way of coding as the way to run code in transactions.

Since the method is still backward compatible we still have the chance
to revert to the old way of handling transactions if we decide to.

In this checkin, the new method Dexie.vip() and the property
Dexie.currentTransaction is moved to be static instead of bound to db.
This is because they dont need to be bound to db instance since they
work towards Promise-specific data.