Version 0.9.8 alpha with updated tests and samples
This commit contains the following changes:

* New transaction() style fully functional and is now the encouraged way
to use the db.transaction() method. Wiki docs will soon be updated.

* Rewrote all unit tests to use the new transaction style
(db.transaction() without callback arguments since they are no longer
needed). Note that the transaction() method itself is backward
compatible but it does not require the caller to use the provided
arguments anymore. I want to encourage not doing that because in a
future version the callback arguments will be removed. Instead use
db.[table] directly everywhere. The only exception from that is
Version.upgrade() since you are working on old version's tables with
different schemas.

* Rewrote all samples and addons to use the new transaction style.

* Support for sub-transactions. This is a natural consequence that we
need to support if users should be able to write transaction blocks that
should be reusable from within an outer transaction block.

* New unit tests for sub-transactions and error handling in sub
transactions.

* Bugfixes in the upgrade / populate framework

* MultiModifyError renamed to ModifyError (MultiModifyError could sound
like there was some kind of race condition or simultanous modifications
taking place, which is not the case with this exception).

* Bugfix: ModifyError.failedKeys did not always contain the last failed
key.

* Collection.keys(), eachKey(), firstKey() and lastKey() now also works
on primary keys.