Nested transactions derive transaction.source
Nested transaction will by default derive the source property from its
parent transaction. The source property can be used to mark the changes
a originated from a certiain source. It can be used when subscribing to
the db.on('changes') event that Dexie.Observable adds to Dexie. When
changes arrive the subscriber may inspect change.source to see if the
change was initiated by itself and the ignore that change. It is also
used by Dexie.Syncable to mark certain changes as originating from a
remote node so that the same changes are never sent back to that node.

This commit makes a nested transaction derive the source property from
its parent transaction by default. This makes it easier for application
code to know that the source property needs only to be set on parent
transaction and all nested function will derive it no matter whether
they enclose their operations within a nested transaction or not.

Also, lowering the NODE_TIMEOUT from 30 to 20 seconds.