Disallow empty "url" to db.syncable.connect()
Also: Let database be ready immediate if the sync operation is performed
by another node. Will eventually in a future checkin also let it be
ready immediate also when own node makes the sync. It's nice to have
access to DB right away and not having to wait for sync to complete.
After all, that's one of the benefits with a local syncronized copy of
the database.

Also, bugfix in the use of PersistedContext.save() from
finallyCommitAllChanges() where the _syncNodes table wasnt included in
transaction. Made the save() method independant on ongoing transaction
but resuse it if it is compatible. Still dont want to include that table
in finallyCommitAllChanges() because it's not vital for transaction
correctness and it would block pollers and cleanup code.