More clear error message when db.open failed
If you perform an operation that is rejected due to an earlier error
occurred in on('populate'), upgrade or on('ready') the error message
should be clear that your operation is not where the error originated
but the error originated when opening the database.

It could be quire confusing otherwise when debugging the code in the
wrong place.