New use case for db.on('ready')
Now if calling db.on('ready') when db is already open, the subscriber
will be fired immediately. That way, it is now possible to use the event
more like the jquery ready event - no matter if database has opened or
not, it will launch when ready to launch.

This means that the on('ready') event differs from how other events
behave. Will soon update wiki documentation to reflect this.