Bug when dumping existing schemas
For migrating to dexie, the sample "dump existing databases" could be
handy for getting the existing schemas from an indexedDB database from
whatever library or bare-bone.

There were some bugs in the code that parsed existing db schemas that
was fixed in this checkin. These bugs does not have any issue number.