#75 Wrong order of changes Dexie.Observable/Syncable
Wrong order of changes Dexie.Observable/Syncable when using sub
transactions.

Redesigned Dexie.Observable to add the changes in the same tick as the
CREATE/UPDATE/DELETE  operation is taking place instead of waiting for
this.onsuccess. Reason is that transaction may be locked when
this.onsuccess is triggered, resulting in that the change revision is
added later than it should have beem and occur after other changes that
actually occurred later.