Just make it work on Safari #77
Adapts to IndexedDBShim without having to call __useShim().
From now on, if you prefer to use IndexedDBShim before the native
IndexedDB support on Safari, just include the indexedDBShim before
including Dexie.js and it will use it instead of the native indexedDB.

This checkin also makes the unit tests able to opt-in or out of the
IndexedDBShim and the iegap polyfill for missing IDB features in
IE10/IE11.