Bumped version to 1.1
Deprecated Dexie.spawn() and replaced it with Dexie.ignoreTransaction()
which is more descriptive. Main reason is that spawn() could be
associated with Task.js spawn() that has another meaning.  Dexie.spawn()
will still be able to use but will log a warning to console that is has
been deprecated.

Next release will be an 1.1 release because of the deprecation of
Dexie.spawn() as well as the changed behavior when including Dexie with
requirejs - before it would still put itself into window but now it will
act as a real amd module.