Made async() independant on Dexie.Promise
This async() / spawn() solution is Promise implementation agnostic in a
deeper level than Task.js and the existing proposals for ecmascript 7
async/await. The reason is that it adapts to the yielded Promise instead
of always converting it to the browser-built-in Promise, with one
exception: If the generator function does not ever use yield, we have to
encapsulate the return value into a Promise. This commit will use a
standard Promise if that occurs instead of returning a Dexie.Promise.
The reason is just to make it more clear that the implementation of
async() and spawn() in it's whole is not a Dexie-specific solution but a
generic solution that would be good to be adopted by the standard makers
of ES7.