A little bit more defensive to IndexedDBShim issues
Now it's possible to run through all unit tests with the IndexedDBShim
activated on chrome / safari. Still gets many failures. Much of them due
to missing support of autoincremented primary key but there are other
issues with parallell queries also.

Current status when running latest IndexedDBShim with this version of
Dexie on Chrome version 42.0.2311.135 m:
335 assertions of 421 passed, 86 failed.

To activate the shim you must use Safari or Chrome and click the
IndexedDBShim checkbox in the unit test main page
~/test/run-unit-tests.html