Speeded up unit test runs
Instead of deleting the database between all tests, it will now just
clear all tables and re-launch on.populate between the tests.

Some tests had to be rewritten because they relied on fix incremental
primary keys.

A new checkbox in the header will make it possible disable this
optimization in case some test doesnt behave as expected (or to just
make sure the optimization is not fippling the tests)