Further optimization of unit tests
In order to rapidly regression test the lib after a change, fast unit
tests are important. I recently optimized the unit test runner so that
instead of deleting and recreating the database between every test, I
just cleared each table and re-ran the populate event for everyone of
them.

This worked well but I found that it became even faster by snapshoting
the DB after on('populate'), and start each unit test by diffing the
current DB state with the snapshot and take appropriate actions to
restore the initial state. This is what is done now.

To disable this optimization, just uncheck its checkbox in the top of
the unit test runner page. Disabling optimization could be done to
troubleshoot issues with the optimization.

Note: Some tests still require recreation of database so they will still
run slow, but most of the tests are fast running now.