Verify indexedDB incompatibility with Promise
Added a unit test that verifies the incompatibility between indexedDB
and window.Promise in that transactions are prematurely committed when
executing Promise.then(). Changed the test to use Dexie.Promise so that
it doesnt fail.

When using window.Promise, transactions are auto-committed in mozilla
38.0.5 but not with Chrome or Operat (Chrome 43.0.2357.124 m,  Opera
30.0). IE11 has no native Promise and no other browser has been tested.

Keep this test as a reference for detecting when Mozilla or other
browsers (for example Edge) will implement compatibility between Promise
and indexedDB.