Issue #114 mapToClass() prototype mess
I have improved the code completion for mapToClass() and defineClass()
to NOT mess with the prototype or the default constructor of
defineClass(). Now code completion features gained by defining a
structure to defineClass() or mapToClass() is NEVER applied in real
world. Only the IDE (at least Visual Studio 2013/2015) will believe that
the classes has the properties from the structure so that we get
excellent code completion there but never mess with the real data.

Removed bad code that wrote to the given Function's prototype in
mapToClass() (also called from defineClass()). Instead, just fix the
instanceTemplate to comply with given structure.

Also removed the sub-optimation to use Object.setPrototypeOf() instead
of cloning the objects to make it inherit a class. Now we always do
Object.create(proto) and clone its members instead.