WriteableTable.bulkAdd()
In response to issue #90, a bulk add method. Does support
hook('creating') if used. This method takes special care of not keeping
any closure reference to the added objects since there are many users
that have had memory consumption and performance issues when adding lots
of items.

I would in the long term also want to create a generic bulk()  method
that would take an array of operations that could be a mix of
add/put/update/deletes, but that will require a bit more complexity.
Starting with this method to get some feedback on whether it improves
performance as it is intended to do.