CRUD hooks should respect Dexie.currentTransaction
There is no good reason for that crud-hooks are the only exception from
respecting Dexie.currentTransaction so that the hook implementations
dont need to use the provided transaction. Performance is probably no
issue.

This test case is a TDD that fails now but should succeed when that
feature is fixed.