Fixes after unit testing with Dexie.Observable on
Specifically, when Dexie.Observable notifies about changes, the "obj"
and "oldObj" properties went through the reading-hooks (so that they
would be derived from it's mapped class). But since reading hooks is not
only for mapping classes, and specifically made the new unit test fail,
I now removed this. I don't think it is nescessary that the changes need
to provide a correct prototype hierarchy in the change notifications
(db.on('changes').