Included babel in order to make it possible to use es6 classes etc. Now only "npm run watch" works as expected. TODO: "npm run build" needs to be adjusted.
Not including all ES2015 modules in babel because some of them generates bloaty code, such as the typeof module that hooks in every time "typeof x === 'object'" is checked. Even though this expression's behavior may change in es2015, we don't need to utilize it like that. We still know what we're doing in typeof expressions.