Issue #120 related - multiple object stores on Safari
Dexie can be used to open databases not created by Dexie. This feature
made Dexie.open() fail in Safari if more than 1 object store due to a
Safari issue when opening a transaction towards multiple object stores.

In this commit, Dexie will silently catch the exception that Safari
would through when opening a  database with multiple stores declared.

It was also an error that in case Dexie.open() fails, the backend
database instance was not set to null, which, (as @dpogue commented)
resulted in that Dexie behaved as if the database was successfully
opened.