Better call stack when IDB throws exception.
Dexie will catch all exceptions and instead reject a promise so that
callers never need to do try..catch for double error handling. However,
the error that IDB throws lacks a call stack in IE and Edge so it can be
horrible to track down where the error occurred. This commit applies a
call stack to errors such as DATA_ERROR when they happen.