Optimized and improved how and when Tables are created. The most important feature with this checkin is the following:
Before, you could not save a Collection and reuse it in another transaction because each Collection and Table where bound to the transaction where they were retrieved. Now, the transaction is looked up when the operation happens instead.

The optimization part of this is that new Transaction() will not have to create new Tables. Instead, there is only one set of Tables. Always the same in transaction and on database.

There are failing unit tests when running with Dexie.Syncable turned on. Those errors occurred in an earlier checkin. Need to find out why.