Removing IncompatiblePromiseError.
IncompatiblePromiseError was introduced to stop people from returning standard Promise from within a transaction scope, such as Promise.all() or async/await native promises.

With the new native promise compatibility, we don't suffer from this anymore. We might warn the user since there are browsers (IE, Safari and Firefox) where indexedDB transactions would fail with TransactionInactiveError if doing this. But we should definitely not throw exception because if using native async functions this will be the case and that is one of the point with this new feature.