Deprecate some Transaction methods.
Deprecated transaction methods and updated samples, addons and unit test code to not use deprecated transaction methods anymore.

Added Debug.deprecated() function that can be used to deprecated any method, function or getter. It will warn on console about the deprecation, with a link to wiki page about the deprecation and a stack trace to find where it's deprecated in the source code.