Put back some deprecated Transaction methods.
The method Transaction.prototype.table() and the property Transaction.prototype.tables has been dependent on in samples and unit tests until recently. Therefore there might be a likeliness of people having copied that code into their apps and it would be worth it to continue support them in the 2.x releases and make them obsolete in far future instead. It's not that big cost to keep there.