Support blobs when hooks are involved (#364)
Support for intrinsic objects (Blobs, Typed arrays ect) in deepClone() so that they will be treated as primitive object and not be deeply cloned. Only deep clone custom objects.

Added unit tests for blob support. The first one succeeds (we had blob support already when not using hooks). But the second one failed before the fix.