Prohibit idling on Promise.race() in transaction. (#379)
If Promises are generated within a transaction scope but not awaited for, these promises were awaited for anyway by the transaction runner using Dexie-specific Promise.follow(). The idea behind this is if users don't code with a Promise-returning style, but just fire off promises without waiting for them or catching them, this should be OK as long as this is being done within a transaction. It can spare the user from having to call Promise.all([]) on everything, which is kind of nice.

However, there might be a risk that user fires off promises that should not be waited for. Specifically, this happens if using Promise.race(). Caller is only interested of waiting for the first promise that completes. So in that case, this behavior could be unwanted.

This commit disables this behavior in case a promise was returned from the scope function, and instead just awaits that promise. This gives the user more control over the situation. Returning a promise is always the recommended way in promise-style coding. So now we only have this behavior if user don't follow the recommended way of returning promises. WIth Dexie that's actually ok, because the promise will be caught by the transaction. But user may choose style and we will now respect a returned Promise as being the sole completion of the transaction.

There are two other places where Promise.follow() is used, and that is in the upgrading framework. I have not adjusted that part of the code in this commit, but it will need to be adjusted the same way - check it a promise is returned, and if so, respect it rather than waiting for Promise.follow() to complete.