Stop using localStorage for recording database names (#399)
Dexie.getDatabaseNames() polyfills indexedDB.getDatabaseNames() for on non-chromium browsers. To do this, it adds the name of every opened table into a localStorage property. Silly not to use an indexedDB database? Reason may have been because I was allergic to storing meta-data in indexedDB. But it's actually even more ridiculous to store it in localStorage so I've moved it to use indexedDB (database '__dbnames') for recording database names so that the feature works in workers as well.

The code size went actually smaller with this change, so I think it's in the right path.

This change will automatically migrate existing localStorage data into the new __dbnames table.