Fix combineUpdateAndUpdate function
Update .mods of previousChange instead the clone of the previousChange object itself.

When the nextChange wanted to update a nested property which was already
set by the previousChange, then we tried to combine with the
previousChange directly instead of its mods.

When the nextChange wanted to update an object in which the
previousChange updated a nested property, we tried to delete the
previous property on the previousChange directly and not its mods.

See also: https://github.com/dfahlander/Dexie.js/issues/397#issuecomment-268009668