Fix resolveConflicts function
Update .mods of clientChange instead of the clientChange object itself.

When resolving server and client UPDATEs we tried to remove the keyPath
from the clientChange object when there was a conflicting path instead of the
clientChange.mods object.

Similarly we tried to remove nested keyPaths which were in conflict from
the clientChange object instead of its .mods object.