Externalize applyChanges and bulkUpdate
applyChanges needs access to the db object. To allow that we use the
globals module in which we save the reference to the database so that
modules needing it can access it.

This is probably not the best of solutions. Maybe after we know what
globals we need to access we can come up with a cleaner solution.