Externalize getOrCreateSyncNode + bug fixes
Have getOrCreateSyncNode and PersistedContext in their own modules.
Add tests for those.

getOrCreateSyncNode bug fixes:

The function in Dexie.Promise.resolve() was never called meaning that we
never set node.myRevision to the currentRevision if initialUpload was
false.

db._changes.lastKey is not a function. Changed that to create a
collection and call lastKey on that.

Add db._changes to the transaction.

The unit test "should set myRevision to the last _changes if
initialUpload is false" makes sure that the Dexie.Promise.resolve
function is called and that node.myRevision is set to the last key in
changes.