Externalize intercomm
Add tests for intercomm in tests-observable-misc.js

Fix bug where db.on('message') was called multiple times with the same
message everytime sendMessage was called multiple times.
Related to: https://github.com/dfahlander/Dexie.js/issues/429

Add some comments in tests-observable-misc.js
Remove unused code. It is incomplete and seems to refer to Syncable and
the sync protocol. Tests for that are in Dexie.Syncable.