Change the format of an intercomm message
Don't merge the contents of the "message" property into the msg before
triggering on('message').

Adjust tests-observable-misc and Dexie.Syncable for that.

Add MessageEvent interface. This is the type received by the
on('message') event.

Remove requestId and isFailure from sendMessage() options. These are
used internally to handle response messages and should not be available
to a consumer.