Fix WhereClause test, hang on upgrade test, and index adjustment.
There were 2 bugs in the previous implementation:

1. Database was not actually being closed in `_open`
2. Object store indices were not properly adjusted to non-Dexie
   index names.

We now properly close the database and discern the object store
applicable to a query using either the name or key, and set the
index name more explicitly on open.

Additionally, a test has been added to check that we can use a
database without the previous version being specified.