Framework for integration tests (#456)
* Added integration test for dexie-relationships
As described in #451.
This is just an example where I use local QUnit tests. Could equally well (and preferrably) invoke a test suite on the external addon instead. But that would require the external addon to release its test suite as part of the npm package and make sure it doesn't bundle dexie with the tests (so we can replace dexie with the local version and run existing tests).
* Removing unused bowser.json files for addons.
* Addons should only depend on dexie as peerDependency