Change options for an existing node
This commit makes two changes.

1. The connect() method was changed to be able to set options for an
existing node. If the node is disconnected then the options will just be
saved on the existing node. If the node is already connected, it will be
disconnected, the options will be saved and the node will be
reconnected. This only happens if the current options for the node
differ from the options connect() was called with.

2. A new method getOptions(url) was added which can be used to get the
current options for the given node.

This commit closes https://github.com/dfahlander/Dexie.js/issues/447

I decided not to implement a setOptions() method as I don't think that
it has much usage if we can have the same effect by using connect().