Partially revert pr 579 (#591)
PR #579 was meant to work around a severe Safari issue with IDBObjectStore.getAll() that when called from a Web worker, results in a page crash followed by an automatic page reload. Since getAll() is just an optimization and not needed for IndexedDB to work, PR 579 shut down this optimization of utilizing on Safari. Due to the nature of the Safari bug (page crash + reload), feature testing was not possible - we could only do user-agent sniffing to avoid the bug.

The user-agent sniffing in PR  #579 did only look for "Safari" in user-agent, which made it also affect Chrome and Opera browsers (which was not the intent). Also, the bug has since been fixed in a later version of Safari + it only happened when called from a web worker, so we needed a more fine grained control of whether the getAll() optimization should be avoided or not. 

This new PR will replace the user-agent code from PR #579 with another more fine grained check that only disables getAll() when:

* browser is Safari (user agent contains "Safari" but not "Chrome" or "Edge")
* we are in a Worker environment
* Safari internal version number is less than 604, as the issue has been fixed in 604.